FeatherViT-Emotion-XXS/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.26.0
onnx>=1.15.0
